+++ doc/whats-new.rst
@@ -25,6 +25,10 @@ Breaking changes
 
 New Features
 ~~~~~~~~~~~~
+- Build :py:meth:`CFTimeIndex.__repr__` explicitly as :py:class:`pandas.Index`. Add ``calendar`` as a new
+  property for :py:class:`CFTimeIndex` and show ``calendar`` and ``length`` in
+  :py:meth:`CFTimeIndex.__repr__` (:issue:`2416`, :pull:`4092`)
+  `Aaron Spring <https://github.com/aaronspring>`_.
 
 
 Bug fixes
@@ -172,7 +176,6 @@ Enhancements
   limited to a reasonable length.
   (:pull:`3905`)
   By `Maximilian Roos <https://github.com/max-sixty>`_
-
 
 Bug fixes
 ~~~~~~~~~